Moved lib from pkg to internal
